give size in meters in myv.points + spherical coordinates
