+++ cookbook/gravmag_euler_classic_expanding_window.py
@@ -60,7 +60,7 @@ for center in centers:
     print "Estimated source location: %s" % (str(results[-1]['point']))
 
 myv.figure()
-myv.points([r['point'] for r in results], size=300.)
+myv.points([r['point'] for r in results], size=100.)
 myv.prisms(model, opacity=0.5)
 axes = myv.axes(myv.outline(bounds), ranges=[b*0.001 for b in bounds])
 myv.wall_bottom(bounds)
